Refactor getTestResultsUseCase.ts and IDBJobsRepository.ts
