+++ server/src/modules/TDDCycles/Domain/IGithubRepository.ts
@@ -1,4 +1,5 @@
 import { CommitDataObject, CommitInformationDataObject } from "./commitInterfaces";
+import { JobDataObject } from "./jobInterfaces";
 
 export interface IGithubRepository {
     getCommits(owner: string, repoName: string): Promise<CommitDataObject[]>;
@@ -7,4 +8,5 @@ export interface IGithubRepository {
     obtainRunsOfGithubActions(owner: string, repoName: string): Promise<any>;
     obtainJobsOfACommit(owner: string, repoName: string, jobId: number, attempt: number): Promise<any>;
     getRunsOfGithubActionsIds(owner: string, repoName: string): Promise<[string, number][]>;
+    getJobsDataFromGithub(owner: string, repoName: string, listOfCommitsWithActions: [string, number][]): Promise<Record<string, JobDataObject>>;
 }
